Remove OxyPlot.SkiaSharp.SvgExporter, replace by OxyPlot.SkiaSharp.SkiaSvgExporter #1532
